Email Web Mail
The JavaMail is an API that is used to compose, write and read electronic messages (emails). The JavaMail API provides a protocol-independent and platform-independent framework for sending and receiving mails. The javax. mail and javax. mail. activation packages contain the core classes of JavaMail API. The Email Web is developed using HTML5, CSS3, BOOTSTRAP4, SERVLETS, JAVA 8, and Tomcat as a server for deployment.